Moves files around
Minor modification to the dataset creation script, `match_datasets.r`, cleaning and creation of some of the variables occurs in the end of this file now.

`analysis.r`now just contains the OLS estimation.

`matching-functional.Rmd` is an R markdown PSM implementation for various treatments
A detailed output file `matching-functional.hmd` is provided for the PSM output.